Splitted test cases for ObjectPropertyNaming and TopLevelPropertyNaming
